requirements.txt
wheel
cmake>=3.24.1
pynrrd>=0.4.2
dipy>=1.4.0
pyyaml>=5.3.1
nibabel>=3.2.1
simpleitk>=2.1.1
reportlab>=3.6.6
pypdf2>=1.26.0
markdown>=3.3.6
xhtml2pdf>=0.2.7
pandas>=1.4.3
flask
flask_cors
flask_jwt_extended
opencv-python
numpy
fury
dmri-amico

